requirements.txt
# Core dependencies
requests>=2.28.0
psutil>=5.9.0
python-dotenv>=0.19.0

# Optional but recommended
numpy>=1.21.0  # For numerical experiments
matplotlib>=3.5.0  # For visualization

# Note: The system will work with just the core dependencies
# Uncomment the optional dependencies if you want to run
# more advanced experiments